[VyGD3IoX] apoc.hashing.fingerprint lossy conversion (neo4j/apoc#378)

+++ core/src/main/java/apoc/hashing/Fingerprinting.java
@@ -330,7 +330,7 @@ public class Fingerprinting {
                 sb.append(d);
             }
         } else if (value instanceof long[]) {
-            for (double l : (long[]) value) {
+            for (long l : (long[]) value) {
                 sb.append(l);
             }
         } else if (value instanceof byte[]) {
